fix: 스타일 컴포넌트 최상단으로 배치
