+++ src/components/Common/FacebookLogin.tsx
@@ -1,41 +1,6 @@
-import ImgSprite from "components/Common/LoginSprite";
+import ImageSprite from "components/Common/LoginSprite";
 import styled from "styled-components";
 import Button from "styles/UI/Button/Button";
-
-export default function FacebookLogin({ bgColor, color }: UI.ButtonProps) {
-    return (
-        <FacebookButtonContainer>
-            <Button bgColor={bgColor} color={color}>
-                {bgColor ? (
-                    <ImgSprite
-                        width={backgroundWhiteFacebook.width}
-                        height={backgroundWhiteFacebook.height}
-                        position={backgroundWhiteFacebook.position}
-                    />
-                ) : (
-                    <ImgSprite
-                        width={backgroundBlueFacebook.width}
-                        height={backgroundBlueFacebook.height}
-                        position={backgroundBlueFacebook.position}
-                    />
-                )}
-                Facebook으로 로그인
-            </Button>
-        </FacebookButtonContainer>
-    );
-}
-
-const backgroundWhiteFacebook: Login.ImageProps = {
-    width: 16,
-    height: 16,
-    position: `-414px -259px`,
-};
-
-const backgroundBlueFacebook: Login.ImageProps = {
-    width: 16,
-    height: 16,
-    position: `-414px -300px`,
-};
 
 const FacebookButtonContainer = styled.div`
     margin: 8px 40px;
@@ -52,3 +17,38 @@ const FacebookButtonContainer = styled.div`
         }
     }
 `;
+
+const backgroundWhiteFacebook: Login.ImageProps = {
+    width: 16,
+    height: 16,
+    position: `-414px -259px`,
+};
+
+const backgroundBlueFacebook: Login.ImageProps = {
+    width: 16,
+    height: 16,
+    position: `-414px -300px`,
+};
+
+export default function FacebookLogin({ bgColor, color }: UI.ButtonProps) {
+    return (
+        <FacebookButtonContainer>
+            <Button bgColor={bgColor} color={color}>
+                {bgColor ? (
+                    <ImageSprite
+                        width={backgroundWhiteFacebook.width}
+                        height={backgroundWhiteFacebook.height}
+                        position={backgroundWhiteFacebook.position}
+                    />
+                ) : (
+                    <ImageSprite
+                        width={backgroundBlueFacebook.width}
+                        height={backgroundBlueFacebook.height}
+                        position={backgroundBlueFacebook.position}
+                    />
+                )}
+                Facebook으로 로그인
+            </Button>
+        </FacebookButtonContainer>
+    );
+}
