+++ src/pages/Login/index.tsx
@@ -2,6 +2,11 @@ import styled from "styled-components";
 import { Contents } from "../../components/Login/Contents";
 import { Footer } from "../../components/Login/Footer/Footer";
 
+const Container = styled.section`
+    display: flex;
+    flex-direction: column;
+`;
+
 export default function Login() {
     return (
         <Container>
@@ -10,9 +15,3 @@ export default function Login() {
         </Container>
     );
 }
-
-// style
-const Container = styled.section`
-    display: flex;
-    flex-direction: column;
-`;
